projeto-01/my_project/alunos/api/urls.py
from django.urls import path,include
from rest_framework import routers
from .views import AlunoViewSet

router = routers.DefaultRouter()

router.register('alunos',AlunoViewSet)

urlpatterns = [
    path('',include(router.urls))
]